backend/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
alembic==1.13.1
# PostgreSQL
asyncpg==0.29.0
psycopg2-binary==2.9.9
# MySQL
aiomysql==0.2.0
pymysql==1.1.0
cryptography==42.0.2

# Redis
redis==5.0.1
celery==5.3.6

# AI / RAG
llama-index==0.10.12
openai==1.12.0

# Document Parsing
unstructured[all-docs]==0.12.4
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.0
pdfplumber==0.10.4

# Vector Database (Local FAISS)
faiss-cpu==1.7.4
numpy>=1.24.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Utils
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.26.0
loguru==0.7.2
